+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/rest/scriptingapi/ScriptingApiModule.java
@@ -28,6 +28,7 @@ public class ScriptingApiModule extends ViewsModule {
     @Override
     protected void configure() {
         addSystemRestResource(ScriptingApiResource.class);
+        bind(ScriptingApiServiceImpl.class).to(ScriptingApiServiceImpl.class).asEagerSingleton();
         jerseyAdditionalComponentsBinder().addBinding().toInstance(TabularResponseWriter.class);
 
         final Multibinder<FieldDecorator> fieldDecoratorBinder = Multibinder.newSetBinder(binder(), FieldDecorator.class);
