+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/rest/scriptingapi/ScriptingApiResource.java
@@ -16,7 +16,6 @@
  */
 package org.graylog.plugins.views.search.rest.scriptingapi;
 
-import com.google.common.eventbus.EventBus;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -34,26 +33,19 @@ import jakarta.ws.rs.core.MediaType;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.graylog.plugins.views.search.Search;
 import org.graylog.plugins.views.search.SearchJob;
-import org.graylog.plugins.views.search.engine.SearchExecutor;
-import org.graylog.plugins.views.search.events.SearchJobExecutionEvent;
 import org.graylog.plugins.views.search.permissions.SearchUser;
 import org.graylog.plugins.views.search.rest.ExecutionState;
 import org.graylog.plugins.views.search.rest.scriptingapi.mapping.AggregationTabularResponseCreator;
-import org.graylog.plugins.views.search.rest.scriptingapi.mapping.MessagesTabularResponseCreator;
 import org.graylog.plugins.views.search.rest.scriptingapi.mapping.QueryFailedException;
 import org.graylog.plugins.views.search.rest.scriptingapi.mapping.QueryParamsToFullRequestSpecificationMapper;
-import org.graylog.plugins.views.search.rest.scriptingapi.mapping.SearchRequestSpecToSearchMapper;
 import org.graylog.plugins.views.search.rest.scriptingapi.request.AggregationRequestSpec;
 import org.graylog.plugins.views.search.rest.scriptingapi.request.MessagesRequestSpec;
 import org.graylog.plugins.views.search.rest.scriptingapi.response.TabularResponse;
 import org.graylog.plugins.views.search.searchtypes.pivot.SortSpec;
 import org.graylog2.audit.jersey.NoAuditEvent;
-import org.graylog2.plugin.database.users.User;
 import org.graylog2.plugin.rest.PluginRestResource;
 import org.graylog2.shared.rest.resources.RestResource;
 import org.graylog2.shared.utilities.StringUtils;
-import org.joda.time.DateTime;
-import org.joda.time.DateTimeZone;
 
 import java.util.List;
 import java.util.Set;
@@ -66,26 +58,13 @@ import static org.graylog2.shared.utilities.StringUtils.splitByComma;
 @Consumes({MediaType.APPLICATION_JSON})
 @RequiresAuthentication
 public class ScriptingApiResource extends RestResource implements PluginRestResource {
-
-    private final SearchExecutor searchExecutor;
-    private final EventBus serverEventBus;
-    private final SearchRequestSpecToSearchMapper searchCreator;
-    private final AggregationTabularResponseCreator aggregationTabularResponseCreator;
-    private final MessagesTabularResponseCreator messagesTabularResponseCreator;
+    private final ScriptingApiService service;
     private final QueryParamsToFullRequestSpecificationMapper queryParamsToFullRequestSpecificationMapper;
 
     @Inject
-    public ScriptingApiResource(final SearchExecutor searchExecutor,
-                                final EventBus serverEventBus,
-                                final SearchRequestSpecToSearchMapper searchCreator,
-                                final AggregationTabularResponseCreator aggregationTabularResponseCreator,
-                                final MessagesTabularResponseCreator messagesTabularResponseCreator,
+    public ScriptingApiResource(final ScriptingApiService service,
                                 final QueryParamsToFullRequestSpecificationMapper queryParamsToFullRequestSpecificationMapper) {
-        this.searchExecutor = searchExecutor;
-        this.serverEventBus = serverEventBus;
-        this.searchCreator = searchCreator;
-        this.aggregationTabularResponseCreator = aggregationTabularResponseCreator;
-        this.messagesTabularResponseCreator = messagesTabularResponseCreator;
+        this.service = service;
         this.queryParamsToFullRequestSpecificationMapper = queryParamsToFullRequestSpecificationMapper;
     }
 
@@ -97,18 +76,8 @@ public class ScriptingApiResource extends RestResource implements PluginRestReso
     @NoAuditEvent("Creating audit event manually in method body.")
     public TabularResponse executeQuery(@ApiParam(name = "queryRequestSpec") @Valid MessagesRequestSpec messagesRequestSpec,
                                         @Context SearchUser searchUser) {
-
         try {
-            //Step 1: map simple request to more complex search
-            Search search = searchCreator.mapToSearch(messagesRequestSpec, searchUser);
-
-            //Step 2: execute search as we usually do
-            final SearchJob searchJob = searchExecutor.executeSync(search, searchUser, ExecutionState.empty());
-            postAuditEvent(searchJob, searchUser.getUser());
-
-            //Step 3: take complex response and try to map it to simpler, tabular form
-            return messagesTabularResponseCreator.mapToResponse(messagesRequestSpec, searchJob, searchUser, getSubject());
-
+            return service.executeQuery(messagesRequestSpec, searchUser, getSubject());
         } catch (IllegalArgumentException | ValidationException | QueryFailedException ex) {
             throw new BadRequestException(ex.getMessage(), ex);
         }
@@ -139,8 +108,8 @@ public class ScriptingApiResource extends RestResource implements PluginRestReso
                     sortOrder,
                     from,
                     size);
-            return executeQuery(messagesRequestSpec, searchUser);
-        } catch (IllegalArgumentException ex) {
+            return service.executeQuery(messagesRequestSpec, searchUser, getSubject());
+        } catch (IllegalArgumentException | QueryFailedException ex) {
             throw new BadRequestException(ex.getMessage(), ex);
         }
     }
@@ -154,15 +123,7 @@ public class ScriptingApiResource extends RestResource implements PluginRestReso
     public TabularResponse executeQuery(@ApiParam(name = "searchRequestSpec") @Valid AggregationRequestSpec aggregationRequestSpec,
                                         @Context SearchUser searchUser) {
         try {
-            //Step 1: map simple request to more complex search
-            Search search = searchCreator.mapToSearch(aggregationRequestSpec, searchUser);
-
-            //Step 2: execute search as we usually do
-            final SearchJob searchJob = searchExecutor.executeSync(search, searchUser, ExecutionState.empty());
-            postAuditEvent(searchJob, searchUser.getUser());
-
-            //Step 3: take complex response and try to map it to simpler, tabular form
-            return aggregationTabularResponseCreator.mapToResponse(aggregationRequestSpec, searchJob, searchUser);
+            return service.executeAggregation(aggregationRequestSpec, searchUser);
         } catch (IllegalArgumentException | ValidationException | QueryFailedException ex) {
             throw new BadRequestException(ex.getMessage(), ex);
         }
@@ -188,14 +149,9 @@ public class ScriptingApiResource extends RestResource implements PluginRestReso
                     splitByComma(groups),
                     splitByComma(metrics)
             );
-            return executeQuery(aggregationRequestSpec, searchUser);
-        } catch (IllegalArgumentException ex) {
+            return service.executeAggregation(aggregationRequestSpec, searchUser);
+        } catch (IllegalArgumentException | QueryFailedException ex) {
             throw new BadRequestException(ex.getMessage(), ex);
         }
-    }
-
-    private void postAuditEvent(SearchJob searchJob, User user) {
-        final SearchJobExecutionEvent searchJobExecutionEvent = SearchJobExecutionEvent.create(user, searchJob, DateTime.now(DateTimeZone.UTC));
-        this.serverEventBus.post(searchJobExecutionEvent);
     }
 }
